requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
hyperbrowser
Pylette
python-multipart
passlib[bcrypt]
httpx
docling
rq-dashboard
supabase
psycopg2
langchain-openai 
langchain-experimental 
langchain-text-splitters
langfuse
openrouter
instructor